requirements-doctool.txt
jupyter-book
sphinx
IPython
numpydoc
sphinx-rtd-theme
pyment
sphinx_codeautolink
